Github Issue #5 - get wheel building again
Summary:
See T202442658. setup.py was not correctly building the shared objects because it was making ft_core.c into a shared object not compiling it into the other shared objects. Fixing this showed up an issue building with ref_debug. Then it made no sense not having a way to run all the tests so I added in installing all the tests in a tests and made an run_all. Also, benchmark_utils was suffering from name space collision with  the concurrent builtin so fixed that too.

With this we should be good to go with the wheel and testing the wheel from now on.

Reviewed By: alexmalyshev

Differential Revision: D63206339

fbshipit-source-id: 9a308657f38166d5c48487d6114003b2e8d621a7