Add general back-compat to ft_utils
Summary:
This diff adds backwards compatibility methods for all of the pyatomics and refactors the headers in ft_utils to split out stuff ft_utils needs to work and that which another library might want to use for backward compat. Specifically we have ft_compat.h which does the heavy lifting and test_compat.py which tests all the behaviors.

A final not that one of the fence methods is just missing from the beta version of ftpython we use internally so I have #if'd around that is a stable way but that code could be removed one day probably.

Reviewed By: alexmalyshev

Differential Revision: D64703872

fbshipit-source-id: f7aedab8f8119102126fd0eca7bdf2e4e8e62215