Adding weave for advance thread support features, initially tls destructors
Summary:
Thread support for high performance features like thread local storage and destructors, cpu affinity and other such things is important in FTPython in a way it just is not in GILPython. Weave is the support library for some of these features.

In this diff we are adding thread local destructors. The snag with this is FTPython is a package and its parts of shared objects. We do not want a shared objects in some other module (say Pillow) to have a C ABI dependency on symbols in weave. The solution here is to provide a header source coupling where intercommunication between modules is done via the python interpreter instead.

This is a bit unusual but then this is for a high performance feature, not yee old scripting.

Reviewed By: alexmalyshev

Differential Revision: D66512938

fbshipit-source-id: a24a1ac958113f2ce68f0f006cd820ee473fc51d