Windows TLS
Summary:
Windows TLS was fundamentally flawed as it mixed FLS and TLS concepts. On windows we only have destructors in FLS. Fortunately, any thread can have fiber support added to it and still remain a thread (it is just and add on) so we can use FLS and destructors just like in posix but on windows. This patch gives us this.

In working on this I found that it was really hard to build ft_utils on windows from github so I simplified this and improved the docs a little at the same time.

Reviewed By: alexmalyshev

Differential Revision: D66974792

fbshipit-source-id: aa51b76ffdabb83b37c49f7f4fba48e71af2bbaf