Fix ft_util macOS GitHub action building
Summary: Previously the load path was not fully using the homebrew-installed packages, meaning it was using system packages. This led to conflicts and the build system not being able to find things like tkinter. This commit changes the load path to be able to take advantage of the installed packages.

Reviewed By: SonicField

Differential Revision: D67682231

fbshipit-source-id: c5ce99b078242c2e7233d06dcdadec75c5ff1cdc