"Fix" ft_utils build for Windows
Summary: ft_utils CI is currently failing for Windows. One reason is that the Windows parser cannot handle a double `;` in the struct declaration. Another reason is that test_weave.c is failing for unknown reasons. We can fix the first one quickly, and for the second one we are temporarily turning off the test_weave.c test until we have time to fix it.

Reviewed By: SonicField

Differential Revision: D67950521

fbshipit-source-id: 1608efcba0482bf9f0e0faad08bd402b96673511