ft_utils building for Windows 32-bit
Summary:
Previously, we were crashing on Windows 32-bit because the process was hanging on to the FLS callback (assumed, it was difficult to pinpoint).

Now, instead of using FLS we are using TLS and hooking into DllMain to get notifications when the thread and process exit. We keep a linked list of requested TLS keys. The linked list is walked on thread detach to call all of the destructors in each of the threads. It is also walked on process exit to free all of the allocated keys.

Doing it this was makes it work on 32-bit windows, as well as keeping our current capabilities. It also has the benefit of not leaking the FLS indices, which previously were not being freed.

Reviewed By: czardoz

Differential Revision: D68632202

fbshipit-source-id: cab95f80cbbbd608c35408b81f82b986f1cc66ee